pyproject.toml
[project]
name = "oversight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "docling>=2.53.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "langchain-core>=0.3.76",
    "langchain-openai>=0.3.33",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "tqdm>=4.67.1",
    "wheel>=0.45.1",
    "FlagEmbedding>=1.3.5",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.2",
    "spacy>=3.6.1",
    "python-igraph>=0.11.8",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
